+++ src/components/Banner/Logo/Logo.jsx
@@ -1,8 +1,11 @@
+import { Link } from "react-router-dom"
 import $HeaderLogo from './Logo.styles.jsx';
 import logo from '../../../assets/carbon-tracker-logo.svg';
 
 export default function Logo() {
     return (
-        <$HeaderLogo src={logo} alt="Carbon Tracker"/>
+        <Link to="/home">
+            <$HeaderLogo src={logo} alt="Carbon Tracker"/>
+        </Link>
     );
 }