+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Implement logic to handle extended  by LDAP calls and objects
 * Create login page and login handling using base auth
 * Handle location/path when manually changed
+* Add Link to home pae on carbon tracker logo
 
 0.36 (2024-10-19)
 ===============